Move Internal headers to Sources/ file and run ./scripts/change_headers.swift.
