+++ FirebaseAppDistribution/Sources/FIRAppDistributionMachOSlice.h
