+++ FirebaseAppDistribution/Sources/FIRAppDistributionUIService.h
@@ -16,7 +16,7 @@
 #import <Foundation/Foundation.h>
 #import <SafariServices/SafariServices.h>
 #import <UIKit/UIKit.h>
-#import "FIRAppDistribution+Private.h"
+#import "FirebaseAppDistribution/Sources/Private/FIRAppDistribution.h"
 
 NS_ASSUME_NONNULL_BEGIN
 /// An instance of this class provides UI elements required for the App Distribution tester
