+++ FirebaseAppDistribution/Sources/FIRAppDistributionUIService.m
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "FIRAppDistributionUIService.h"
+#import "FirebaseAppDistribution/Sources/FIRAppDistributionUIService.h"
+#import "FirebaseAppDistribution/Sources/FIRFADLogger.h"
+#import "FirebaseCore/Sources/Private/FirebaseCoreInternal.h"
+#import "FirebaseAppDistribution/Sources/Public/FIRAppDistribution.h"
+
 #import <AuthenticationServices/AuthenticationServices.h>
 #import <SafariServices/SafariServices.h>
 #import <UIKit/UIKit.h>
-#import "FIRAppDistribution.h"
-#import "FIRFADLogger+Private.h"
-#import "FirebaseCore/Sources/Private/FirebaseCoreInternal.h"
 
 @implementation FIRAppDistributionUIService
