+++ FirebaseAppDistribution/Sources/Public/FirebaseAppDistribution.h
@@ -12,5 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "FIRAppDistribution.h"
-#import "FIRAppDistributionRelease.h"
+#import "FirebaseAppDistribution/Sources/Public/FIRAppDistribution.h"
+#import "FirebaseAppDistribution/Sources/Public/FIRAppDistributionRelease.h"
