+++ FirebaseAppDistribution/Tests/Unit/FIRAppDistributionMachOTests.m
@@ -15,8 +15,7 @@
 #import <Foundation/Foundation.h>
 #import <XCTest/XCTest.h>
 
-#import "FIRAppDistributionMachO+Private.h"
-#import "FirebaseAppDistribution/FIRAppDistributionMachO+Private.h"
+#import "FirebaseAppDistribution/Sources/FIRAppDistributionMachO.h"
 #import "FirebaseCore/Sources/Private/FirebaseCoreInternal.h"
 
 @interface FIRAppDistributionMachOTests : XCTestCase
