+++ Package.swift
@@ -40,6 +40,10 @@ let package = Package(
       targets: ["FirebaseAuth"]
     ),
     .library(
+      name: "FirebaseAppDistribution",
+      targets: ["FirebaseAppDistribution"]
+    )
+    .library(
       name: "FirebaseCrashlytics",
       targets: ["FirebaseCrashlytics"]
     ),
@@ -97,6 +101,7 @@ let package = Package(
       dependencies: [
         "FirebaseAuth",
         "FirebaseABTesting",
+        "FirebaseAppDistribution",
         "FirebaseFunctions",
         "Firebase",
         "FirebaseCrashlytics",
@@ -317,6 +322,28 @@ let package = Package(
         "FIRAuthTests.m",
         "FIRUserTests.m",
       ],
+      cSettings: [
+        .headerSearchPath("../../.."),
+      ]
+    ),
+    .target(
+      name: "FirebaseAppDistribution",
+      dependencies: ["FirebaseCore",
+                     "FirebaseInstallations",
+                     "GoogleUtilities_UserDefaults",
+                     "GoogleUtilities_AppDelegateSwizzler",
+                     "GoogleDataTransport"],
+      path: "FirebaseAppDistribution/Sources",
+      publicHeadersPath: "Public",
+      cSettings: [
+        .headerSearchPath("../../"),
+        .define("FIRAppDistribution_VERSION", to: "0.0.1"), // TODO: Fix version
+      ]
+    ),
+    .testTarget(
+      name: "AppDistributionUnit",
+      dependencies: ["FirebaseAppDistribution", "OCMock"],
+      path: "FirebaseAppDistribution/Tests/Unit",
       cSettings: [
         .headerSearchPath("../../.."),
       ]
